requirements.txt
# --- START OF FILE requirements.txt (Pinned Versions) ---
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain==0.1.16
langchain-community==0.0.32
langchain-openai==0.1.3
pinecone-client==3.2.2  # A version that still worked with the old LangChain syntax
trafilatura
langdetect
deep-translator
pytest